Improved error handling
The SVG parsing code was exit() ing when there was an error parsing the
SVG data. Now, if the SVG root-level elements are not correct, no paths
will be created. If an individual path has problems, that path will not
be created, but any well-formed SVG paths will be created.
We need some tests to run against this code to verify it works under
all conditions.